+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Reorder the fields on the new report form per design
 - Show if a report for ISPF is ODA or non-ODA on the report show view
 - Show if a report for ISPF is ODA or non-ODA on the report edit form
+- Upload templates for ISPF reports use the ODA type of the report
 
 ## Release 139 - 2023-11-14